rp: Create built-in RP implementation.
