+++ VisualPinball.Unity/VisualPinball.Unity/Rendering.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df41a046588ec2945af6656bab83a61b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
