+++ VisualPinball.Unity/VisualPinball.Unity/Rendering/RenderPipeline.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0600de79d8e20bc41bad183a799279ec
+guid: 72a7be43d14c0b64da79f0d24d29d3a8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
