+++ VisualPinball.Unity/VisualPinball.Unity/Rendering/StandardRenderPipeline.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2923f6bbc1e2284095db7ae86717a56
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
